requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
docutils==0.21.2
idna==3.10
Kivy==2.3.0
Kivy-Garden==0.1.5
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
